requirements.txt
requests>=2.31.0
agno>=1.7.0
python-dotenv>=1.0.0
openai>=1.0.0

# Google Sheets integration (optional)
google-api-python-client>=2.100.0
google-auth>=2.20.0
google-auth-oauthlib>=1.0.0
google-auth-httplib2>=0.1.0

# Email notifications
secure-smtplib>=0.1.1
